refactor: move types to `app/types.ts`
